Homepage Frontend Changes
Changes some of the alignments on the homepage and modified some more CSS to make it more accessible and welcoming. I also introduced some buttons that can link the user to our social media sites to learn more about us and show some support. These still need to be made functional but I still need to go through the process of creating the pages and what not.